feat: migrate workspace to nx 22.x
migrate nx workspace to 22.x and update plugins to reference 22.x devkit

BREAKING CHANGE: You must migrate your nx workspace to 22.x to use this version of the @nxlv/python
plugin

re #338

+++ packages/data-migration/src/migrator/dynamodb/migration.ts
@@ -528,7 +528,7 @@ export abstract class DynamoDBMigrationBase extends MigrationBase {
           Role: lambdaOptions.iamRoleArn,
           Handler: 'handler.handler',
           Code: {
-            ZipFile: fs.readFileSync(zipPath),
+            ZipFile: fs.readFileSync(zipPath) as Uint8Array,
           },
           Description: `Lambda function for migration ${this.namespace}/${this.name}/${this.version} stream`,
           Timeout: lambdaOptions.timeout,
